requirements.txt
python-telegram-bot
selenium
webdriver-manager
pytz
beautifulsoup4 
